Fixup git operations (#6)
- Fix remote logic
- Account for new lock error
- Set committer details when updating reflog

This fixes some issues when dealing with git index repos managed by
cargo that the internal testing in this crate didn't discover, namely
that cargo (via git2) does not set remotes, and updating reflogs
requires that a committer be set by gix which isn't the case in CI
environments (I guess that doesn't apply if the remote is a directory on
the same filesystem?). So this PR resolves both those issues.