Add concurrent helpers to `Async/RemoteSparseIndex` (#11)
- Update gix
- Add concurrent/parallel helpers
- Ship parsing/disk ops to threadpool

This came up during https://github.com/rustsec/rustsec/pull/923, but it
is often more efficient to send many concurrent requests with an async
runtime than to use the blocking reqwest client with a threadpool eg
rayon. This just adds `krates` helpers to request many crates
concurrently for both the blocking and async implementations so that end
users don't have to do it themselves.