Make locking behavior configurable (#14)
### Checklist

* [x] I have read the [Contributor Guide](../../CONTRIBUTING.md)
* [x] I have read and agree to the [Code of
Conduct](../../CODE_OF_CONDUCT.md)
* [x] I have added a description of my changes and why I'd like them
included in the section below

### Description of Changes

Makes the locking behavior configurable by API users. 

Needed for `cargo audit` to avoid inexplicable hangs when waiting on a
lock: it will first attempt to lock the directory immediately, and if
that fails, print a message and only after that start waiting for the
lock.

### Related Issues

#12, https://github.com/rustsec/rustsec/pull/934