update `gix` to the latest version (#27)
Please note that I have removed the `max-performance-safe` feature
which pushes the choice to the consumer of (any) library that uses
`gix`. It's also causing less C to be compiled by default.

### Checklist

* [x] I have read the [Contributor Guide](../../CONTRIBUTING.md)
* [x] I have read and agree to the [Code of
Conduct](../../CODE_OF_CONDUCT.md)
* [x] I have added a description of my changes and why I'd like them
included in the section below

### Description of Changes

An upgrade to the latest `gix`, which includes less dependencies now and
compiles a little faster
because of it. It also avoids making decisions about performance (and
the inclusion of C-crates)
by removing the `max-performance-safe` feature.

The motivation for this update is to get a new `tame-index` release for
consumption in [this
PR](https://github.com/obi1kenobi/cargo-semver-checks/pull/531)
so that it can also use the latest `gix` version without duplication.

### Related Issues

None

---------

Co-authored-by: Jake Shadle <jake.shadle@embark-studios.com>