Set patch to latest for `windows-targets`
`windows-targets` apparently has breaking changes in patch versions as show in #40, so this just sets the version to the latest that compiles correctly. If they make further breaking chnanges we might need to pin it, but hopefully that won't happen. Another option is just to remove the dependency altogether since it's not really doing anything to be honest.