Fix broken links in the pull request template
Paths are relative to the PR url, not the source source file
per markdown docs, so these were going to the organization
rather than the repo.

This feels like a github issue, since having the links be
relative to the published url isn't very template-like,
but in the mean time this work-around makes the links
clickable.