Align with cargo (#69)
As noted in the comment, this changes the url kind from `u64` -> `usize`
to match cargo, which unfortunately uses the default `derive(Hash)`
which for enums uses `::core::intrinsics::discriminant_value` which for
non-repr enums is...pointer size. This means the hash computed is
different between 32 and 64-bit targets when they could easily be the
same. :(

Resolves:
https://github.com/EmbarkStudios/cargo-deny/issues/540#issuecomment-2262990465